@@ -7,20 +7,14 @@ class AppTheme {
 
   static const TextTheme textTheme = TextTheme();
 
-  static bool isJunat =
-      AppFlavorEnvironment.appFlavor == FlavorEnvironment.junat.name;
+  static ThemeData get light =>
+      _theme(AppColorSchemeJunat.lightScheme());
 
-  static ThemeData get light => _theme(isJunat
-      ? AppColorSchemeJunat.lightScheme()
-      : AppColorScheme.lightScheme());
+          static ThemeData get dark => _theme(
+          AppColorSchemeJunat.darkScheme());
 
-  static ThemeData get dark => _theme(
-        isJunat
-            ? AppColorSchemeJunat.darkScheme()
-            : AppColorScheme.darkScheme(),
-      );
-
-  static ThemeData _theme(ColorScheme colorScheme) => ThemeData(
+  static ThemeData _theme(ColorScheme colorScheme) =>
+      ThemeData(
         useMaterial3: true,
         fontFamily: AppFonts.poppins,
         brightness: colorScheme.brightness,
@@ -30,19 +24,20 @@ class AppTheme {
           displayColor: colorScheme.onSurface,
         ),
         appBarTheme: AppBarTheme(
-          backgroundColor: colorScheme.background,
+          backgroundColor: colorScheme.surface,
+          elevation: .5,
           titleTextStyle: TextStyle(
             fontWeight: FontWeight.bold,
             color: colorScheme.onSurface,
-            fontSize: 24,
+            fontSize: 18,
             fontFamily: AppFonts.poppins,
           ),
         ),
         dividerTheme: DividerThemeData(
-          color: colorScheme.outline.withOpacity(0.6),
+          color: colorScheme.outline.withValues(alpha: 0.6),
           thickness: 1,
         ),
-        scaffoldBackgroundColor: colorScheme.background,
+        scaffoldBackgroundColor: colorScheme.surface,
         canvasColor: colorScheme.surface,
       );
 }
@@ -50,19 +45,25 @@ class AppTheme {
 extension ThemeExtension on BuildContext {
   ThemeData get theme => Theme.of(this);
 
-  TextTheme get textTheme => Theme.of(this).textTheme;
+  TextTheme get textTheme =>
+      Theme
+          .of(this)
+          .textTheme;
 
-  ColorScheme get colorScheme => Theme.of(this).colorScheme;
+  ColorScheme get colorScheme =>
+      Theme
+          .of(this)
+          .colorScheme;
 
   TextStyle get h2 => theme.textTheme.displayMedium!;
 
-  TextStyle get h3 => theme.textTheme.displaySmall!;
+  TextStyle get h3 => theme.textTheme.titleLarge!;
 
   TextStyle get h4 => theme.textTheme.headlineMedium!;
 
-  TextStyle get h5 => theme.textTheme.headlineSmall!;
+  TextStyle get h5 => theme.textTheme.titleLarge!;
 
-  TextStyle get h6 => theme.textTheme.titleLarge!;
+  TextStyle get h6 => theme.textTheme.bodyLarge!;
 
   TextStyle get sub1 => theme.textTheme.titleMedium!;
 
@@ -75,15 +76,23 @@ extension ThemeExtension on BuildContext {
   TextStyle get body2Bold =>
       theme.textTheme.bodyMedium!.copyWith(fontWeight: FontWeight.w800);
 
-  TextStyle get bodyError => theme.textTheme.bodyMedium!
-      .copyWith(color: Theme.of(this).colorScheme.error);
+  TextStyle get bodyError =>
+      theme.textTheme.bodyMedium!
+          .copyWith(color: Theme
+          .of(this)
+          .colorScheme
+          .error);
 
   TextStyle get caption => theme.textTheme.bodySmall!;
 
   TextStyle get smallest => theme.textTheme.bodyMedium!;
 
-  TextStyle get captionError => theme.textTheme.bodySmall!
-      .copyWith(color: Theme.of(this).colorScheme.error);
+  TextStyle get captionError =>
+      theme.textTheme.bodySmall!
+          .copyWith(color: Theme
+          .of(this)
+          .colorScheme
+          .error);
 
   TextStyle get button => theme.textTheme.labelLarge!;
 
@@ -91,10 +100,13 @@ extension ThemeExtension on BuildContext {
 
   TextStyle get overline => theme.textTheme.labelSmall!;
 
-  TextStyle get appBarTitle => theme.textTheme.bodyLarge!.copyWith(
+  TextStyle get appBarTitle =>
+      theme.textTheme.bodyLarge!.copyWith(
         fontWeight: FontWeight.w600,
       );
 
   bool get isDarkMode =>
-      MediaQuery.of(this).platformBrightness == Brightness.dark;
+      MediaQuery
+          .of(this)
+          .platformBrightness == Brightness.dark;
 }
