@@ -21,210 +21,176 @@ class DashboardScreen extends GetView<DashboardController> {
   Widget build(BuildContext context) {
     controller.loadDependencies();
     return Scaffold(
-        appBar: AppBar(
-          centerTitle: false,
-          title: SizedBox(
-            width: context.width,
-            height: kToolbarHeight,
-            child: FittedBox(
-              fit: BoxFit.scaleDown,
-              alignment: Alignment.centerLeft,
-              child: Obx(
-                () => Text(
-                  AppFlavorEnvironment.appFlavor ==
-                          FlavorEnvironment.automanager.name
-                      ? controller.company.value.name ?? 'Auto Manager'
-                      : controller.junatCompanyName.value,
-                  textAlign: TextAlign.left,
-                ),
+      appBar: AppBar(
+        centerTitle: false,
+        title: SizedBox(
+          width: context.width,
+          // height: kToolbarHeight,
+          child: FittedBox(
+            fit: BoxFit.scaleDown,
+            alignment: Alignment.centerLeft,
+            child: Obx(
+              () => Text(
+                controller.company.value.name ?? 'Auto Manager',
+                textAlign: TextAlign.left,
               ),
             ),
           ),
-          toolbarHeight: kToolbarHeight * 1.3,
-          actions: <Widget>[
-           /* _logoWrapper(
+        ),
+        toolbarHeight: kToolbarHeight * 1.3,
+        actions: <Widget>[
+          Obx(
+            () => _logoWrapper(
               context,
-              child: IconButton(
-                onPressed: () {},
-                icon: const Icon(
-                  Iconsax.notification5,
-                  size: 28,
-                ),
-              ),
-            ),*/
-            if (AppFlavorEnvironment.appFlavor ==
-                FlavorEnvironment.automanager.name)
-              Obx(
-                () => _logoWrapper(
-                  context,
-                  child: controller.company.value.logoUrl != null &&
+              child:
+                  controller.company.value.logoUrl != null &&
                           controller.company.value.logoUrl != ''
                       ? CachedNetworkImage(
-                          imageUrl: controller.company.value.logoUrl!,
-                          errorWidget: (BuildContext context, String url,
-                                  dynamic error) =>
-                              const Icon(Icons.error),
-                        )
+                        imageUrl: controller.company.value.logoUrl!,
+                        errorWidget:
+                            (BuildContext context, String url, dynamic error) =>
+                                const Icon(Icons.error),
+                      )
                       : const AppLogo(),
-                ),
-              )
-            else
-              _logoWrapper(context, child: const AppLogo()),
-          ],
-        ),
-        body: Column(
-          children: <Widget>[
-            const Divider(),
-            Expanded(child: _buildDashboardContents(context)),
-          ],
-        ));
+            ),
+          ),
+        ],
+      ),
+      body: Column(
+        children: <Widget>[
+          Expanded(child: _buildDashboardContents(context)),
+        ],
+      ),
+    );
   }
 
   Widget _buildDashboardContents(BuildContext context) {
     return SingleChildScrollView(
       child: Padding(
         padding: AppPaddings.mA,
-        child: Column(children: <Widget>[
-          _buildDashboardSummaryDateField(context),
-          const AppSpacing(
-            v: 10,
-          ),
-          Obx(
-            () => DashboardSummaryCard(
-              title: 'Revenue',
-              value: DataFormatter.getLocalCompactCurrencyFormatter(context)
-                  .format(
-                controller.dashboardSummary.value.revenue,
+        child: Column(
+          children: <Widget>[
+            _buildDashboardSummaryDateField(context),
+            const AppSpacing(v: 10),
+            Obx(
+              () => DashboardSummaryCard(
+                title: 'Revenue',
+                value: DataFormatter.getLocalCompactCurrencyFormatter(
+                  context,
+                ).format(controller.dashboardSummary.value.revenue),
+                icon: Iconsax.moneys5,
+                onTap: () {},
+                valueIcon:
+                    (controller.dashboardSummary.value.revenue ?? 0) < 0
+                        ? Icon(
+                          Ionicons.arrow_down,
+                          size: 14,
+                          color: context.colorScheme.error,
+                        )
+                        : null,
               ),
-              icon: Iconsax.moneys5,
-              onTap: () {},
-              valueIcon: (controller.dashboardSummary.value.revenue ?? 0) < 0
-                  ? Icon(
-                      Ionicons.arrow_down,
-                      size: 14,
-                      color: context.colorScheme.error,
-                    )
-                  : null,
             ),
-          ),
-          const AppSpacing(
-            v: 10,
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: <Widget>[
-              Expanded(
-                child: Obx(
-                  () => DashboardSummaryCard(
-                    title: 'Sales',
-                    value:
-                        DataFormatter.getLocalCompactCurrencyFormatter(context)
-                            .format(
-                      controller.dashboardSummary.value.sales,
-                    ),
-                    icon: Icons.money_rounded,
-                  ),
-                ),
-              ),
-              const AppSpacing(
-                h: 10,
-              ),
-              Expanded(
-                child: Obx(
-                  () => DashboardSummaryCard(
-                    title: 'Expenses',
-                    value:
-                        DataFormatter.getLocalCompactCurrencyFormatter(context)
-                            .format(
-                      controller.dashboardSummary.value.expenses,
-                    ),
-                    icon: IconlyBold.wallet,
-                  ),
-                ),
-              ),
-            ],
-          ),
-          const AppSpacing(
-            v: 10,
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: <Widget>[
-              Expanded(
-                child: Obx(
-                  () => DashboardSummaryCard(
-                    title: 'Rentals',
-                    value:
-                        DataFormatter.getLocalCompactCurrencyFormatter(context)
-                            .format(
-                      controller.dashboardSummary.value.rentalSales,
-                    ),
-                    icon: Ionicons.car,
-                  ),
-                ),
-              ),
-              const AppSpacing(
-                h: 10,
-              ),
-              Expanded(
-                child: Obx(
-                  () => DashboardSummaryCard(
-                    title: 'Vehicles',
-                    value:
-                        controller.dashboardSummary.value.vehicles.toString(),
-                    icon: Ionicons.speedometer,
-                  ),
-                ),
-              ),
-            ],
-          ),
-          const AppSpacing(
-            v: 10,
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: <Widget>[
-              Expanded(
-                child: Obx(
-                  () => DashboardSummaryCard(
-                    title: 'Drivers',
-                    value: controller.dashboardSummary.value.drivers.toString(),
-                    icon: IconlyBold.discovery,
-                  ),
-                ),
-              ),
-              const AppSpacing(
-                h: 10,
-              ),
-              Expanded(
-                child: Obx(
-                  () => DashboardSummaryCard(
-                    title: 'Customers',
-                    value:
-                        controller.dashboardSummary.value.customers.toString(),
-                    icon: IconlyBold.user_3,
-                  ),
-                ),
-              ),
-            ],
-          ),
-          const AppSpacing(
-            v: 10,
-          ),
-          Card(
-            color: context.colorScheme.outline.withOpacity(0.1),
-            shadowColor: context.colorScheme.outline.withOpacity(0.1),
-            child: Column(
+            const AppSpacing(v: 10),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: <Widget>[
-                Align(
-                  alignment: Alignment.centerRight,
-                  child: _buildMonthDateField(context),
+                Expanded(
+                  child: Obx(
+                    () => DashboardSummaryCard(
+                      title: 'Sales',
+                      value: DataFormatter.getLocalCompactCurrencyFormatter(
+                        context,
+                      ).format(controller.dashboardSummary.value.sales),
+                      icon: Icons.money_rounded,
+                    ),
+                  ),
                 ),
-                _buildWeeklyChart(context),
+                const AppSpacing(h: 10),
+                Expanded(
+                  child: Obx(
+                    () => DashboardSummaryCard(
+                      title: 'Expenses',
+                      value: DataFormatter.getLocalCompactCurrencyFormatter(
+                        context,
+                      ).format(controller.dashboardSummary.value.expenses),
+                      icon: IconlyBold.wallet,
+                    ),
+                  ),
+                ),
               ],
             ),
-          ),
-        ]),
+            const AppSpacing(v: 10),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: <Widget>[
+                Expanded(
+                  child: Obx(
+                    () => DashboardSummaryCard(
+                      title: 'Rentals',
+                      value: DataFormatter.getLocalCompactCurrencyFormatter(
+                        context,
+                      ).format(controller.dashboardSummary.value.rentalSales),
+                      icon: Ionicons.car,
+                    ),
+                  ),
+                ),
+                const AppSpacing(h: 10),
+                Expanded(
+                  child: Obx(
+                    () => DashboardSummaryCard(
+                      title: 'Vehicles',
+                      value:
+                          controller.dashboardSummary.value.vehicles.toString(),
+                      icon: Ionicons.speedometer,
+                    ),
+                  ),
+                ),
+              ],
+            ),
+            const AppSpacing(v: 10),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: <Widget>[
+                Expanded(
+                  child: Obx(
+                    () => DashboardSummaryCard(
+                      title: 'Drivers',
+                      value:
+                          controller.dashboardSummary.value.drivers.toString(),
+                      icon: IconlyBold.discovery,
+                    ),
+                  ),
+                ),
+                const AppSpacing(h: 10),
+                Expanded(
+                  child: Obx(
+                    () => DashboardSummaryCard(
+                      title: 'Customers',
+                      value:
+                          controller.dashboardSummary.value.customers
+                              .toString(),
+                      icon: IconlyBold.user_3,
+                    ),
+                  ),
+                ),
+              ],
+            ),
+            const AppSpacing(v: 10),
+            Card(
+              color: context.colorScheme.outline.withValues(alpha: 0.1),
+              shadowColor: context.colorScheme.outline.withValues(alpha: 0.1),
+              child: Column(
+                children: <Widget>[
+                  Align(
+                    alignment: Alignment.centerRight,
+                    child: _buildMonthDateField(context),
+                  ),
+                  _buildWeeklyChart(context),
+                ],
+              ),
+            ),
+          ],
+        ),
       ),
     );
   }
@@ -237,7 +203,7 @@ class DashboardScreen extends GetView<DashboardController> {
       child: Container(
         padding: AppPaddings.mA,
         decoration: BoxDecoration(
-          color: context.colorScheme.outline.withOpacity(0.2),
+          color: context.colorScheme.outline.withValues(alpha: 0.2),
           borderRadius: AppBorderRadius.card,
         ),
         // padding: AppPaddings.sA.add(AppPaddings.sH),
@@ -245,19 +211,15 @@ class DashboardScreen extends GetView<DashboardController> {
           crossAxisAlignment: CrossAxisAlignment.center,
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            const Icon(
-              IconlyLight.calendar,
-            ),
-            const AppSpacing(
-              h: 10,
-            ),
+            const Icon(IconlyLight.calendar),
+            const AppSpacing(h: 10),
             Obx(
               () => Text(
                 DataFormatter.formatDateToTextMonthYear(
                   controller.selectedMonthYear.value.toIso8601String(),
                 ),
               ),
-            )
+            ),
           ],
         ),
       ),
@@ -274,40 +236,49 @@ class DashboardScreen extends GetView<DashboardController> {
             text: 'Weekly Sales',
             alignment: ChartAlignment.near,
           ),
-          primaryXAxis: const NumericAxis(
-            axisLine: AxisLine(width: 0),
+          primaryXAxis: CategoryAxis(
+            axisLine: const AxisLine(width: 0),
             labelPosition: ChartDataLabelPosition.outside,
-            majorTickLines: MajorTickLines(width: 0),
-            majorGridLines: MajorGridLines(width: 0),
-            labelFormat: 'Week {value}',
+            majorTickLines: const MajorTickLines(width: 0),
+            majorGridLines: const MajorGridLines(width: 0),
+            // labelFormat: 'Week {value}',
+            axisLabelFormatter: (AxisLabelRenderDetails args) {
+              return ChartAxisLabel(
+                'Week ${args.value + 1}',
+                const TextStyle(fontWeight: FontWeight.normal),
+              );
+            },
           ),
+
           primaryYAxis: NumericAxis(
             isVisible: false,
-            numberFormat: NumberFormat.compactCurrency(
-              symbol: '',
-            ),
+            numberFormat: NumberFormat.compactCurrency(symbol: ''),
           ),
           series: _getRoundedColumnSeries(context),
           tooltipBehavior: controller.tooltipBehavior,
         );
-      }
+      },
     );
   }
 
   List<ColumnSeries<ChartData, num>> _getRoundedColumnSeries(
-      BuildContext context) {
+    BuildContext context,
+  ) {
     return <ColumnSeries<ChartData, num>>[
       ColumnSeries<ChartData, num>(
         color: context.colorScheme.outline,
         width: 0.7,
         dataLabelSettings: const DataLabelSettings(
-            isVisible: true, labelAlignment: ChartDataLabelAlignment.top),
+          isVisible: true,
+          labelAlignment: ChartDataLabelAlignment.top,
+        ),
         dataSource: controller.salesForTheMonthData,
         borderRadius: BorderRadius.circular(15),
         xValueMapper: (ChartData sales, _) => sales.xValue,
         yValueMapper: (ChartData sales, _) => sales.yValue,
-        onRendererCreated:
-            (ChartSeriesController<ChartData, num> chartSeriesController) {
+        onRendererCreated: (
+          ChartSeriesController<ChartData, num> chartSeriesController,
+        ) {
           controller.chartSeriesController = chartSeriesController;
         },
       ),
@@ -322,26 +293,23 @@ class DashboardScreen extends GetView<DashboardController> {
       child: Container(
         padding: AppPaddings.mA,
         decoration: BoxDecoration(
-            borderRadius: AppBorderRadius.card,
-            border: Border.all(
-              color: context.colorScheme.secondary.withOpacity(0.4),
-            )),
+          borderRadius: AppBorderRadius.card,
+          border: Border.all(
+            color: context.colorScheme.secondary.withValues(alpha: 0.4),
+          ),
+        ),
         child: Row(
           crossAxisAlignment: CrossAxisAlignment.center,
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            const Icon(
-              IconlyLight.calendar,
-            ),
-            const AppSpacing(
-              h: 10,
-            ),
+            const Icon(IconlyLight.calendar),
+            const AppSpacing(h: 10),
             Obx(
               () => Text(
                 controller.dateText.value,
                 style: const TextStyle(fontWeight: FontWeight.bold),
               ),
-            )
+            ),
           ],
         ),
       ),
@@ -357,10 +325,11 @@ class DashboardScreen extends GetView<DashboardController> {
           height: 50,
           width: 50,
           decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(15),
-              border: Border.all(
-                color: context.colorScheme.secondary.withOpacity(0.4),
-              )),
+            borderRadius: BorderRadius.circular(15),
+            border: Border.all(
+              color: context.colorScheme.secondary.withValues(alpha: 0.4),
+            ),
+          ),
           child: child,
         ),
       ),
