@@ -1,12 +1,14 @@
-import 'package:automanager/core/core.dart';
+
 import 'package:flutter/cupertino.dart';
 
 import '../../../feature/auto_manager/presentation/presentation.dart';
 
+final
+
 List<Widget> mobileNavPages = <Widget>[
- if (UserPermissions.validator.canViewDashboard) const DashboardScreen(),
+const DashboardScreen(),
   const SalesScreen(),
   const ExpensesScreen(),
-  if (UserPermissions.validator.canViewRentals) const RentalScreen(),
+  const RentalScreen(),
   const MoreScreen(),
 ];
