@@ -2,8 +2,6 @@ import 'package:automanager/core/core.dart';
 import 'package:automanager/core/presentation/theme/app_theme.dart';
 import 'package:flutter/material.dart';
 
-import '../../utils/flavor_config.dart';
-import '../utils/utils.dart';
 
 class AppLogo extends StatelessWidget {
   const AppLogo({this.height, this.width, super.key});
@@ -22,11 +20,9 @@ class AppLogo extends StatelessWidget {
         width: width,
       ),
       mobile: Image.asset(
-        AppFlavorEnvironment.appFlavor == FlavorEnvironment.automanager.name
-            ? (context.isDarkMode
+             (context.isDarkMode
                 ? AssetImages.appLogoWhite
-                : AssetImages.appLogoBlack)
-            : AssetImages.junatLogo,
+                : AssetImages.appLogoBlack),
         height: height,
         width: width,
       ),
