@@ -47,8 +47,8 @@ extension EnvironmentX on Environment {
   String get url {
     return <Environment, String>{
       Environment.development:
-        'https://auto-manager-api-6yxs.onrender.com/api/',
-      Environment.production: 'https://auto-manager-api-6yxs.onrender.com/api/',
+      'http://192.168.10.216:3000/api/',
+      Environment.production: 'https://auto-manager-api.vercel.app/api/',
     }[this]!;
   }
 }
