@@ -1,8 +1,6 @@
 import 'package:automanager/core/presentation/theme/app_theme.dart';
 import 'package:automanager/feature/auto_manager/presentation/expense/expense.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
 import 'package:get/get.dart';
 import 'package:iconly/iconly.dart';
@@ -99,7 +97,7 @@ class ExpensesScreen extends GetView<ExpenseController> {
                           Visibility(
                             visible: UserPermissions.validator.canUpdateExpense,
                             child: SlidableAction(
-                              backgroundColor: context.colorScheme.background,
+                              backgroundColor: context.colorScheme.surface,
                               //  foregroundColor: Colors.red,
                               icon: IconlyLight.edit,
                               label: 'Edit',
@@ -111,7 +109,7 @@ class ExpensesScreen extends GetView<ExpenseController> {
                           Visibility(
                             visible: UserPermissions.validator.canDeleteExpense,
                             child: SlidableAction(
-                              backgroundColor: context.colorScheme.background,
+                              backgroundColor: context.colorScheme.surface,
                               foregroundColor: Colors.red,
                               icon: IconlyLight.delete,
                               label: 'Delete',
@@ -224,8 +222,9 @@ class ExpensesScreen extends GetView<ExpenseController> {
                         Text(
                           expense.expenseId,
                           textAlign: TextAlign.left,
-                          style: const TextStyle(
+                          style: context.sub2.copyWith(
                             fontWeight: FontWeight.bold,
+                            fontSize: 12,
                           ),
                         ),
                         Text(DataFormatter.formatDate(expense.createdAt ?? '')),
@@ -274,7 +273,9 @@ class ExpensesScreen extends GetView<ExpenseController> {
             ),
             Expanded(
               child: Center(
-                child: Text('Amount(GHS)'),
+                child: Text('Amount(GHS)',
+                  overflow: TextOverflow.ellipsis,
+                  maxLines: 1,),
               ),
             ),
             /* Expanded(
@@ -299,13 +300,16 @@ class ExpensesScreen extends GetView<ExpenseController> {
               controller.totalAmount.value,
             ),
             textAlign: TextAlign.left,
-            style: Theme.of(context).textTheme.titleLarge?.copyWith(
+            style: Theme.of(context).textTheme.titleSmall?.copyWith(
                   fontWeight: FontWeight.bold,
                 ),
           ),
         ),
         Chip(
-          backgroundColor: context.colorScheme.background,
+          backgroundColor: context.colorScheme.surface,
+          labelStyle: context.caption.copyWith(
+            color: context.colorScheme.onSurface,
+          ),
           label: Obx(
             () => Text(
               controller.dateText.value,
